client chagnes updated

+++ admin-api/database/migrations/tenant/2020_07_06_112633_create_table_mission_donation_impact.php
@@ -16,7 +16,7 @@ class CreateTableMissionDonationImpact extends Migration
         Schema::create('mission_impact_donation', function (Blueprint $table) {
             $table->uuid('mission_impact_donation_id')->primary();
             $table->unsignedBigInteger('mission_id');
-            $table->float('amount', 16, 4);
+            $table->decimal('amount', 16, 4);
             $table->timestamps();
             $table->softDeletes();
